experimental_scripts/requirements.txt
os
numpy
pandas
scikit-learn
tkinter
tqdm
natsort
opencv-python
Pillow
mss
scipy
tifffile